requirements.txt
streamlit
deepface
opencv-python-headless
numpy
pillow
tensorflow
tf-keras

